Workaround lockups with HD Audio controller
Turning the PCIe controller on before the HD audio controller is turned on
can lead to hung tee processes, hung kernel tasks and other badness.

If HDA_CONTROLLER_BUS_ID is set we can turn the HD Audio Controller on
before the PCIe controller is turned on.
This can cause a PCI error:
>can't change power state from D3cold to D0 (config space inaccessible)
But that's better than hanging the system.

When turning off the card the order doesn't seem to matter.
But we do turn on Powersave for the HDA controller again,
as the nvidia card can't fully power off otherwise.